updated single time form

+++ db/migrate/20170627191712_create_work_segments.rb
@@ -5,7 +5,7 @@ class CreateWorkSegments < ActiveRecord::Migration[5.1]
       t.integer :workday_id
       t.string :task
       t.integer :project_id
-      t.integer :timeEntry
+      t.decimal :timeEntry
       t.string :segmentNotes
 
       t.timestamps
